Fixed absolute URLs with not root base
When the base URL includes some default path, the root URLs in a response now
should be handled properly.  Some weird comparison magic to match the paths
before building the `Route` object.  Excuse the mess with the `Link` header
stuff, couldn't finish it on the train, will wrap it up soon.