Added query string support for GET
GET requests now also put the params into the URL query string.  Added a test
for this scenario.  Also found some bugs in the actual URL requesting (used the
`sample.py` to check this).  Need to run this more often and figure out a way
to achieve this via a unit test.