Add in test server for integration tests
Using the `multiprocessing` library, a simple test HTTP server is spun
up in another process so that actual requests can be made.  The data
returned is meant to be a summary of the request information.  Only a
few tests exist for now, but more can be added as additional things are
discovered/requested.