URL handling overhaul
In an attempt to ease some of the URL handling logic, this is an attempt
to remove all of the string concatenation tricks in URL building into a
set of handlers in the `restler.url` space.  This allows for easier
traversal from parsing and some logic in attribute based building.

Additional things were some clean up with the Makefile, migrating some
of the hairier blocks of setup into builders, centralizing path
splitting and cleaning, a centralized builder/lookup wrapper to allow
for re-using route objects, not building routes layer by layer but just
a single step jump, cleaned up query inheritance and switched to using
the proper parsers as much as possible, and removed the wild dictionary
combination logic for param combinations.