Documentation overhaul
In preparation of defining sphinx docs for the project, most of the
class and method definitions have had docstrings updated or added to
provide RST based documentation and provide an easy to follow
documentation on how they work and should be used.  As with all
documentation, these are very much a work in progress.